+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - The `conflicts_with` fields to describe conflicting packages in the `package.toml` metadata file. Two conflicting packages cannot be symlinked at the same time.
 - The `size` field in the source fields in the `targets.toml` metadata file.
 - Deprecation information fields (deprecated_from, disabled_from and reason), to allow specifying deprecation and disabling dates of packages or specific versions.
+- The `--overwrite` flag to the link command to overwrite existing links from another package that are conflicting.
 
 ### Changes
 - The build system now includes a new `package-build` xtask to create a full Packit prebuild for the release.